+++ src/math.ts
@@ -15,31 +15,9 @@ import { PowPipe } from './math/pow.pipe';
 export * from './math/bytes.pipe';
 export * from './math/ceil.pipe';
 export * from './math/floor.pipe';
-export * from './math/max.pipe';
-export * from './math/mean.pipe';
-export * from './math/min.pipe';
 export * from './math/round.pipe';
-export * from './math/sum.pipe';
 export * from './math/degrees.pipe';
 export * from './math/radians.pipe';
 export * from './math/random.pipe';
 export * from './math/sqrt.pipe';
 export * from './math/pow.pipe';
-
-
-
-export const NG2_MATH_PIPES = [
-    BytesPipe,
-    CeilPipe,
-    FloorPipe,
-    MaxPipe,
-    MeanPipe,
-    MinPipe,
-    RoundPipe,
-    SumPipe,
-    DegreesPipe,
-    RadiansPipe,
-    RandomPipe,
-    SqrtPipe,
-    PowPipe
-];