+++ src/string.ts
@@ -1,20 +1,3 @@
-import { LeftPadPipe } from './string/left-pad.pipe';
-import { MatchPipe } from './string/match.pipe';
-import { PadPipe } from './string/pad.pipe';
-import { ReplacePipe } from './string/replace.pipe';
-import { RightPadPipe } from './string/right-pad.pipe';
-import { SplitPipe } from './string/split.pipe';
-import { TestPipe } from './string/test.pipe';
-import { TrimPipe } from './string/trim.pipe';
-import { NewlinesPipe } from './string/newlines.pipe';
-import { CapitalizePipe } from './string/capitalize.pipe';
-import { UpperFirstPipe } from './string/upperfirst.pipe';
-import { TemplatePipe } from './string/template.pipe';
-import { EncodeURIPipe } from './string/encode-uri.pipe';
-import { EncodeURIComponentPipe } from './string/encode-uri-component.pipe';
-import { TruncatePipe } from './string/truncate.pipe';
-import { RepeatPipe } from './string/repeat.pipe';
-
 export * from './string/left-pad.pipe';
 export * from './string/match.pipe';
 export * from './string/pad.pipe';
@@ -31,24 +14,3 @@ export * from './string/encode-uri.pipe';
 export * from './string/encode-uri-component.pipe';
 export * from './string/truncate.pipe';
 export * from './string/repeat.pipe';
-
-
-
-export const NG2_STRING_PIPES = [
-    LeftPadPipe,
-    MatchPipe,
-    PadPipe,
-    ReplacePipe,
-    RightPadPipe,
-    SplitPipe,
-    TestPipe,
-    TrimPipe,
-    NewlinesPipe,
-    CapitalizePipe,
-    UpperFirstPipe,
-    TemplatePipe,
-    EncodeURIPipe,
-    EncodeURIComponentPipe,
-    TruncatePipe,
-    RepeatPipe
-];